requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
decorator==5.1.1
entrypoints==0.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
rich==13.3.2
scrapeghost==0.4.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.20.0
structlog==22.3.0
tiktoken==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
yarl==1.8.2
zipp==3.15.0
